requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --pip-args='--cert /etc/ssl/certs/ca-certificates.crt'
#
absl-py==2.3.1
    # via
    #   ml-collections
    #   open-spiel
annotated-types==0.7.0
    # via pydantic
asttokens==3.0.1
    # via stack-data
attrs==25.4.0
    # via open-spiel
beautifulsoup4==4.14.3
    # via strategicwm (pyproject.toml)
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
comm==0.2.3
    # via ipywidgets
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via
    #   ipython
    #   retry
etils==1.13.0
    # via strategicwm (pyproject.toml)
executing==2.2.1
    # via stack-data
fonttools==4.61.0
    # via matplotlib
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via
    #   google-generativeai
    #   strategicwm (pyproject.toml)
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via strategicwm (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
ipython==9.8.0
    # via
    #   ipywidgets
    #   pyvis
    #   strategicwm (pyproject.toml)
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.8
    # via strategicwm (pyproject.toml)
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via pyvis
jsonpickle==4.1.1
    # via pyvis
jupyterlab-widgets==3.0.16
    # via ipywidgets
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.7
    # via strategicwm (pyproject.toml)
matplotlib-inline==0.2.1
    # via ipython
ml-collections==1.1.0
    # via open-spiel
networkx==3.6
    # via
    #   pyvis
    #   strategicwm (pyproject.toml)
numpy==2.3.5
    # via
    #   contourpy
    #   matplotlib
    #   open-spiel
    #   scipy
    #   strategicwm (pyproject.toml)
open-spiel==1.6.9
    # via strategicwm (pyproject.toml)
packaging==25.0
    # via matplotlib
parso==0.8.5
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==12.0.0
    # via matplotlib
prompt-toolkit==3.0.52
    # via ipython
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
py==1.11.0
    # via retry
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   google-generativeai
    #   strategicwm (pyproject.toml)
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.2.5
    # via
    #   httplib2
    #   matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyvis==0.3.2
    # via strategicwm (pyproject.toml)
pyyaml==6.0.3
    # via ml-collections
requests==2.32.5
    # via google-api-core
retry==0.9.2
    # via strategicwm (pyproject.toml)
rsa==4.9.1
    # via google-auth
scipy==1.16.3
    # via open-spiel
six==1.17.0
    # via python-dateutil
soupsieve==2.8
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
tqdm==4.67.1
    # via
    #   google-generativeai
    #   strategicwm (pyproject.toml)
traitlets==5.14.3
    # via
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
wcwidth==0.2.14
    # via prompt-toolkit
widgetsnbextension==4.0.15
    # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# pip
